[FE][WL] move helper functions to utils folder
