+++ src/utils/interaction-helper.ts
